requirements.txt
aiohttp==2.3.10
arrow==0.12.1
async-timeout==2.0.1
asyncio==3.4.3
boto3==1.9.4
botocore==1.12.4
chardet==3.0.4
Click==7.0
docutils==0.14
guizero==0.6.3
idna==2.7
idna-ssl==1.1.0
jmespath==0.9.3
multidict==4.4.2
Pillow==5.2.0
PyMata==2.18
pymata-aio==2.28
pyserial==3.4
python-dateutil==2.7.3
s3transfer==0.1.13
scrython==1.5.0
six==1.11.0
urllib3==1.23
websockets==7.0
yarl==1.2.0
